+++ src/issuingToken/issuingToken.php
@@ -16,6 +16,7 @@ class IssuingToken extends Resource
     public $walletId;
     public $walletName;
     public $merchantId;
+    public $url;
     public $externalId;
     public $tags;
     public $status;
@@ -41,6 +42,7 @@ class IssuingToken extends Resource
         - walletId [string]: wallet provider which the token is bounded to. ex: "google"
         - walletName [string]: wallet name. ex: "GOOGLE"
         - merchantId [string]: merchant unique id. ex: "5656565656565656"
+        - url [string]: token URL returned by the API. ex: "https://example.com/issuing-token/5656565656565656"
 
     ## Attributes (IssuingToken only):
         - id [string]: unique id returned when IssuingToken is created. ex: "5656565656565656"
@@ -69,11 +71,12 @@ class IssuingToken extends Resource
         $this->walletId = Checks::checkParam($params, "walletId");
         $this->walletName = Checks::checkParam($params, "walletName");
         $this->merchantId = Checks::checkParam($params, "merchantId");
+        $this->url = Checks::checkParam($params, "url");
         $this->externalId = Checks::checkParam($params, "externalId");
         $this->tags = Checks::checkParam($params, "tags");
         $this->status = Checks::checkParam($params, "status");
-        $this->created = Checks::checkParam($params, "created");
-        $this->updated = Checks::checkParam($params, "updated");
+        $this->created = Checks::checkDateTime(Checks::checkParam($params, "created"));
+        $this->updated = Checks::checkDateTime(Checks::checkParam($params, "updated"));
         $this->activationCode = Checks::checkParam($params, "activationCode");
         $this->methodCode = Checks::checkParam($params, "methodCode");
         $this->deviceType = Checks::checkParam($params, "deviceType");
@@ -178,7 +181,7 @@ class IssuingToken extends Resource
     ## Return:
         - target IssuingToken with updated attributes
      */
-    public static function update($id, $params, $user = null)
+    public static function update($id, $params = [], $user = null)
     {
         return Rest::patchId($user, IssuingToken::resource(), $id, $params);
     }