+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Changed
+- url attribute to IssuingToken resource
 
 ## [0.16.0] - 2026-06-01
 ### Added